Update handleclient_write_patched_unit_test.py
to support the error in python 3.9
'AssertionError: {'val[17 chars]1, 'type': 'URL', 'data': {'format': 'string',[276 chars]x'}]} != {'val[17 chars]1, 'ttl': 86400, 'type': 'URL', 'timestamp': '[276 chars]>'}]}

Diff is 800 characters long. Set self.maxDiff to None to see it. : The PUT request payload that the method "add_additional_URL" assembled differs from the expected. This does not necessarily mean that it is wrong, it might just be a different way to talking to the Handle Server. Please run an integration test to check this and update the exptected PUT request accordingly.'